Restructured the FitNesse-tests.

+++ specifications/src/main/java/se/inera/webcert/spec/TaBortFraga.groovy
@@ -1,22 +1,15 @@
 package se.inera.webcert.spec
 
 import se.inera.webcert.spec.util.RestClientFixture
-/**
- * Created by pehr on 9/23/13.
- */
-public class TaBortFraga extends RestClientFixture {
+
+class TaBortFraga extends RestClientFixture {
 
     String internReferens
     String externReferens
 
-    public void execute() {
+    def execute() {
         def restClient = createRestClient(baseUrl)
-
-        if (internReferens)
-            restClient.delete(path: "questions/${internReferens}")
-
-        if (externReferens)
-            restClient.delete(path: "questions/extern/${externReferens}")
+        if (internReferens) restClient.delete(path: "questions/${internReferens}")
+        if (externReferens) restClient.delete(path: "questions/extern/${externReferens}")
     }
 }
-
